requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.22.0
jinja2>=3.1.3
python-multipart>=0.0.9
openpyxl>=3.1.5
pydantic>=1.10.13
itsdangerous>=2.1.2
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
authlib>=1.6.0
httpx>=0.28.0
cryptography>=46.0.0
